260309_personalized_recommendation/backend/app/agents/requirements.txt
# Core dependencies
semantic-kernel>=1.0.0
pydantic>=2.0.0

# Azure SDK
azure-search-documents>=11.4.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
azure-core>=1.29.0

# Data processing
pandas>=2.0.0
pyarrow>=14.0.0

# For Azure Blob Storage with abfs protocol
adlfs>=2023.12.0

# Jupyter support (for image rendering)
ipython>=8.0.0
